backend/requirements.txt
alembic
fastapi
psycopg2
pydantic
python-dotenv
sqlmodel
uvicorn